+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 # Revision history for ollama-haskell
 
-## 0.1.1.3 -- 2024-11-06
+## 0.1.1.3 -- 2024-11-08
 
+* Increase response timeout to 15 minutes
+* Added encodeImage utility function that converts image filePath to base64 image data.
 * Added generateJson and chatJson. High level function to return response in Haskell type.
 
 ## 0.1.0.3 -- 2024-11-05